requirements.txt
disnake==2.9.1
python-dotenv
openai